Extract `JPEG{,Un}StuffedByteStreamGenerator` out of `BitStreamerJPEGBenchmark`
